pyproject.toml
[tool.poetry]
name = "eware"
version = "0.1.0"
description = ""
authors = ["Evgen4567 <aes1993rich@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
arrow = "^1.2.2"
python-dotenv = "^0.20.0"
loguru = "^0.6.0"
httpx = "^0.23.0"
pandas = "^1.4.2"
emoji = "^1.7.0"
aiogram = "^2.20"
pydantic = "^1.9.1"
openpyxl = "^3.0.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
